Give up on SHA1 embedding.
We can only embed the SHA1 of the blob, not the commit that includes it. And I
can't find a way to link to a blob by *its* SHA1, rather than the SHA1 of the
commit that includes it. (Previous commit was broken.) So, we could embed the
SHA1 of the corresponding git blob in the file as a piece of text, but that
really doesn't buy you much. Doesn't seem worth the trouble to me.